shared dto to llm conversation between chat and evaluation
